IMPALA-5002: Associate toolchain build scripts/flags with built packages
Adds the current git hash of the native-toolchain so that
the build scripts (including things like compilation/linker
flags) can be associated with the packages later. Before
this change, it was difficult to know if the way in which a
package was built changed by only looking at the packages
downloaded by Impala.

Change-Id: Ida6f1eee0c20de7fbfdbb716e5c7ac8686b558b6